Unify `BaseEndpoint`
